updated permissions for buttons, .bootstrapTable when modifying user's authorized user rights, set default selection for new document orgs to user's admin orgs

+++ app/assets/javascripts/dashboard/documents.js.coffee
@@ -23,30 +23,33 @@ $ ->
   # DOCUMENTS LISTENERS BEGIN
 
   $(document).on 'click', '#document-new', ->
-    data = 'protocol_id': $(this).data('protocol-id')
-    $.ajax
-      type: 'GET'
-      url: '/dashboard/documents/new'
-      data: data
+    if $(this).data('permission')
+      data = 'protocol_id': $(this).data('protocol-id')
+      $.ajax
+        type: 'GET'
+        url: '/dashboard/documents/new'
+        data: data
 
   $(document).on 'click', '.document-edit', ->
-    row_index   = $(this).parents('tr').data('index')
-    document_id = $(this).parents('table#documents-table').bootstrapTable('getData')[row_index].id
-    data = 'protocol_id': $('#document-new').data('protocol-id')
-    $.ajax
-      type: 'GET'
-      url: "/dashboard/documents/#{document_id}/edit"
-      data: data
+    if $(this).data('permission')
+      row_index   = $(this).parents('tr').data('index')
+      document_id = $(this).parents('table#documents-table').bootstrapTable('getData')[row_index].id
+      data = 'protocol_id': $('#document-new').data('protocol-id')
+      $.ajax
+        type: 'GET'
+        url: "/dashboard/documents/#{document_id}/edit"
+        data: data
 
   $(document).on 'click', '.document-delete', ->
-    row_index   = $(this).parents('tr').data('index')
-    document_id = $(this).parents('table#documents-table').bootstrapTable('getData')[row_index].id
-    data = 'protocol_id': $('#document-new').data('protocol-id')
-    if confirm "Are you sure you want to delete the selected Document from this Protocol?"
-      $.ajax
-        type: 'DELETE'
-        url: "/dashboard/documents/#{document_id}"
-        data: data
+    if $(this).data('permission')
+      row_index   = $(this).parents('tr').data('index')
+      document_id = $(this).parents('table#documents-table').bootstrapTable('getData')[row_index].id
+      data = 'protocol_id': $('#document-new').data('protocol-id')
+      if confirm "Are you sure you want to delete the selected Document from this Protocol?"
+        $.ajax
+          type: 'DELETE'
+          url: "/dashboard/documents/#{document_id}"
+          data: data
 
   $(document).on 'change', '#document-doc-type', ->
     if $(this).val() == 'other'
